+++ src/core/CommonShaders.cpp
@@ -300,7 +300,7 @@ R"(
 	void ParticleMatcapShader::draw(const std::initializer_list<const Mesh*>& meshes,
 		VirtualCamera& camera, const box2px& bounds, const box2px& viewport, float radius) {
 		glEnable(GL_SCISSOR_TEST);
-		glScissor(bounds.position.x, viewport.dimensions.y - bounds.position.y - bounds.dimensions.y, bounds.dimensions.x, bounds.dimensions.y);
+		glScissor((int)bounds.position.x, (int)(viewport.dimensions.y - bounds.position.y - bounds.dimensions.y), (int)(bounds.dimensions.x), (int)(bounds.dimensions.y));
 		begin().set("MIN_DEPTH", camera.getNearPlane()).set("matcapTexture", matcapTexture, 0).set("MAX_DEPTH",
 			camera.getFarPlane()).set("RADIUS", radius).set("bounds", bounds).set("viewport", viewport).set(camera,
 				viewport).set("PoseMat", float4x4::identity()).draw(
@@ -329,13 +329,13 @@ R"(
 	void ParticleMatcapShader::draw(const std::list<const Mesh*>& meshes,
 		VirtualCamera& camera, const box2px& bounds, const box2px& viewport, float radius) {
 		glEnable(GL_SCISSOR_TEST);
-		glScissor(bounds.position.x, viewport.dimensions.y - bounds.position.y - bounds.dimensions.y, bounds.dimensions.x, bounds.dimensions.y);
+		glScissor((int)bounds.position.x, (int)(viewport.dimensions.y - bounds.position.y - bounds.dimensions.y), (int)(bounds.dimensions.x), (int)(bounds.dimensions.y));
 
 		begin().set("MIN_DEPTH", camera.getNearPlane()).set("matcapTexture", matcapTexture, 0).set("MAX_DEPTH",
 			camera.getFarPlane()).set("RADIUS", radius).set("bounds", bounds).set("viewport", viewport).set("PoseMat",
 				float4x4::identity()).set(camera, viewport).draw(
 					meshes, GLMesh::PrimitiveType::POINTS).end();
-		glScissor(viewport.position.x, viewport.position.x, viewport.dimensions.x, viewport.dimensions.y);
+		glScissor((int)viewport.position.x, (int)viewport.position.x, (int)viewport.dimensions.x, (int)viewport.dimensions.y);
 		glDisable(GL_SCISSOR_TEST);
 	}
 	void ParticleMatcapShader::draw(
@@ -352,7 +352,7 @@ R"(
 				GLMesh::PrimitiveType::POINTS);
 		}
 		end();
-		glScissor(viewport.position.x, viewport.position.x, viewport.dimensions.x, viewport.dimensions.y);
+		glScissor((int)viewport.position.x, (int)viewport.position.x, (int)viewport.dimensions.x, (int)viewport.dimensions.y);
 		glDisable(GL_SCISSOR_TEST);
 	}
 	ParticleIdShader::ParticleIdShader(
